GO-2318 Tidy Middleware and carry out func
