+++ core/relations.go
@@ -4,14 +4,9 @@ import (
 	"context"
 	"errors"
 
-	"github.com/gogo/protobuf/types"
-
 	"github.com/anyproto/anytype-heart/core/block"
-	"github.com/anyproto/anytype-heart/core/block/object/objectcreator"
 	"github.com/anyproto/anytype-heart/core/domain"
 	"github.com/anyproto/anytype-heart/pb"
-	"github.com/anyproto/anytype-heart/pkg/lib/bundle"
-	"github.com/anyproto/anytype-heart/util/pbtypes"
 )
 
 func (mw *Middleware) ObjectTypeRelationAdd(cctx context.Context, req *pb.RpcObjectTypeRelationAddRequest) *pb.RpcObjectTypeRelationAddResponse {
@@ -50,89 +45,6 @@ func (mw *Middleware) ObjectTypeRelationRemove(cctx context.Context, req *pb.Rpc
 			Description: getErrorDescription(err),
 		},
 	}
-}
-
-func (mw *Middleware) ObjectCreateObjectType(cctx context.Context, req *pb.RpcObjectCreateObjectTypeRequest) *pb.RpcObjectCreateObjectTypeResponse {
-	response := func(code pb.RpcObjectCreateObjectTypeResponseErrorCode, id string, details *types.Struct, err error) *pb.RpcObjectCreateObjectTypeResponse {
-		m := &pb.RpcObjectCreateObjectTypeResponse{ObjectId: id, Details: details, Error: &pb.RpcObjectCreateObjectTypeResponseError{Code: code}}
-		if err != nil {
-			m.Error.Description = err.Error()
-		}
-		return m
-	}
-
-	creator := getService[objectcreator.Service](mw)
-	createReq := objectcreator.CreateObjectRequest{
-		ObjectTypeKey: bundle.TypeKeyObjectType,
-		InternalFlags: req.InternalFlags,
-		Details:       req.Details,
-	}
-	id, newDetails, err := creator.CreateObject(cctx, req.SpaceId, createReq)
-	if err != nil {
-		return response(pb.RpcObjectCreateObjectTypeResponseError_UNKNOWN_ERROR, "", nil, err)
-	}
-
-	return response(pb.RpcObjectCreateObjectTypeResponseError_NULL, id, newDetails, nil)
-}
-
-func (mw *Middleware) ObjectCreateRelation(cctx context.Context, req *pb.RpcObjectCreateRelationRequest) *pb.RpcObjectCreateRelationResponse {
-	response := func(id string, object *types.Struct, err error) *pb.RpcObjectCreateRelationResponse {
-		if err != nil {
-			return &pb.RpcObjectCreateRelationResponse{
-				Error: &pb.RpcObjectCreateRelationResponseError{
-					Code:        pb.RpcObjectCreateRelationResponseError_UNKNOWN_ERROR,
-					Description: err.Error(),
-				},
-			}
-		}
-		key := pbtypes.GetString(object, bundle.RelationKeyRelationKey.String())
-		return &pb.RpcObjectCreateRelationResponse{
-			Error: &pb.RpcObjectCreateRelationResponseError{
-				Code: pb.RpcObjectCreateRelationResponseError_NULL,
-			},
-			ObjectId: id,
-			Key:      key,
-			Details:  object,
-		}
-	}
-	creator := getService[objectcreator.Service](mw)
-	createReq := objectcreator.CreateObjectRequest{
-		ObjectTypeKey: bundle.TypeKeyRelation,
-		Details:       req.Details,
-	}
-	id, newDetails, err := creator.CreateObject(cctx, req.SpaceId, createReq)
-	if err != nil {
-		return response("", nil, err)
-	}
-	return response(id, newDetails, err)
-}
-
-func (mw *Middleware) ObjectCreateRelationOption(cctx context.Context, req *pb.RpcObjectCreateRelationOptionRequest) *pb.RpcObjectCreateRelationOptionResponse {
-	response := func(id string, newDetails *types.Struct, err error) *pb.RpcObjectCreateRelationOptionResponse {
-		if err != nil {
-			return &pb.RpcObjectCreateRelationOptionResponse{
-				Error: &pb.RpcObjectCreateRelationOptionResponseError{
-					Code:        pb.RpcObjectCreateRelationOptionResponseError_UNKNOWN_ERROR,
-					Description: err.Error(),
-				},
-			}
-		}
-		return &pb.RpcObjectCreateRelationOptionResponse{
-			Error: &pb.RpcObjectCreateRelationOptionResponseError{
-				Code: pb.RpcObjectCreateRelationOptionResponseError_NULL,
-			},
-			ObjectId: id,
-			Details:  newDetails,
-		}
-	}
-
-	creator := getService[objectcreator.Service](mw)
-	createReq := objectcreator.CreateObjectRequest{
-		ObjectTypeKey: bundle.TypeKeyRelationOption,
-		Details:       req.Details,
-	}
-	id, newDetails, err := creator.CreateObject(cctx, req.SpaceId, createReq)
-	return response(id, newDetails, err)
 }
 
 func (mw *Middleware) RelationListRemoveOption(cctx context.Context, request *pb.RpcRelationListRemoveOptionRequest) *pb.RpcRelationListRemoveOptionResponse {
